fix: extrat type to its own type and impl conversion
