+++ .config/dictionaries/project.dic
@@ -257,6 +257,7 @@ ureq
 userid
 utimensat
 UTXO
+uuidv
 vitss
 Vkey
 vkeywitness
