+++ rust/rbac-registration/src/cardano/cip509/types/mod.rs
@@ -1,0 +1,6 @@
+//! Types use in CIP-509
+
+pub mod cert_key_hash;
+pub mod ed25519_pubkey;
+pub mod tx_input_hash;
+pub mod uuidv4;
